backend/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
chromadb>=0.5.0
langchain-chroma>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
pypdf>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
ruff>=0.8.0
